face_recognition/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dlib @ file:///C:/Users/HP%20ELITEBOOK/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
numpy==2.2.6
opencv-python==4.13.0.92
pillow==12.1.1
scikit-learn==1.7.2
scipy==1.15.3
threadpoolctl==3.6.0
Werkzeug==3.1.5
